Move the existing `esp-wifi` package into a directory
