+++ esp-wifi/examples/static_ip.rs
