+++ esp-wifi/smoketest.bat
