+++ esp-wifi/src/common_adapter/common_adapter_esp32.rs
