+++ esp-wifi/src/common_adapter/phy_init_data_esp32s2.rs
