+++ esp-wifi/src/compat/common.rs
